fix request login / user can login both with email and username, fix config security , fix data lombok
