+++ Golden Ticket New Back-End/goldenmovieticketnew/src/main/java/com/example/goldenmovieticketnew/dtos/MovieDto.java
@@ -1,11 +1,141 @@
 package com.example.goldenmovieticketnew.dtos;
 
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import java.time.LocalDate;
 
-@Data
 public class MovieDto {
+    public String getId() {
+        return id;
+    }
+    public MovieDto(){
+
+    }
+    public MovieDto(String name, String smallImageURl, String shortDescription, String longDescription, String largeImageURL, String director, String actors, String categories, LocalDate releaseDate, int duration, String trailerURL, String language, String rated) {
+        this.name = name;
+        this.smallImageURl = smallImageURl;
+        this.shortDescription = shortDescription;
+        this.longDescription = longDescription;
+        this.largeImageURL = largeImageURL;
+        this.director = director;
+        this.actors = actors;
+        this.categories = categories;
+        this.releaseDate = releaseDate;
+        this.duration = duration;
+        this.trailerURL = trailerURL;
+        this.language = language;
+        this.rated = rated;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getSmallImageURl() {
+        return smallImageURl;
+    }
+
+    public void setSmallImageURl(String smallImageURl) {
+        this.smallImageURl = smallImageURl;
+    }
+
+    public String getShortDescription() {
+        return shortDescription;
+    }
+
+    public void setShortDescription(String shortDescription) {
+        this.shortDescription = shortDescription;
+    }
+
+    public String getLongDescription() {
+        return longDescription;
+    }
+
+    public void setLongDescription(String longDescription) {
+        this.longDescription = longDescription;
+    }
+
+    public String getLargeImageURL() {
+        return largeImageURL;
+    }
+
+    public void setLargeImageURL(String largeImageURL) {
+        this.largeImageURL = largeImageURL;
+    }
+
+    public String getDirector() {
+        return director;
+    }
+
+    public void setDirector(String director) {
+        this.director = director;
+    }
+
+    public String getActors() {
+        return actors;
+    }
+
+    public void setActors(String actors) {
+        this.actors = actors;
+    }
+
+    public String getCategories() {
+        return categories;
+    }
+
+    public void setCategories(String categories) {
+        this.categories = categories;
+    }
+
+    public LocalDate getReleaseDate() {
+        return releaseDate;
+    }
+
+    public void setReleaseDate(LocalDate releaseDate) {
+        this.releaseDate = releaseDate;
+    }
+
+    public int getDuration() {
+        return duration;
+    }
+
+    public void setDuration(int duration) {
+        this.duration = duration;
+    }
+
+    public String getTrailerURL() {
+        return trailerURL;
+    }
+
+    public void setTrailerURL(String trailerURL) {
+        this.trailerURL = trailerURL;
+    }
+
+    public String getLanguage() {
+        return language;
+    }
+
+    public void setLanguage(String language) {
+        this.language = language;
+    }
+
+    public String getRated() {
+        return rated;
+    }
+
+    public void setRated(String rated) {
+        this.rated = rated;
+    }
+
     private String id;
     private String name;
     private String smallImageURl;
