+++ Golden Ticket New Back-End/goldenmovieticketnew/src/main/java/com/example/goldenmovieticketnew/models/Branch.java
@@ -10,47 +10,9 @@ import org.springframework.data.mongodb.core.mapping.Document;
 
 @Document(collection = "branchs")
 @NoArgsConstructor
-
+@Getter
+@Setter
 public class Branch {
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getImgURL() {
-        return imgURL;
-    }
-
-    public void setImgURL(String imgURL) {
-        this.imgURL = imgURL;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getPhoneNo() {
-        return phoneNo;
-    }
-
-    public void setPhoneNo(String phoneNo) {
-        this.phoneNo = phoneNo;
-    }
 
     @Id
     private String id;
