+++ Golden Ticket New Back-End/goldenmovieticketnew/src/main/java/com/example/goldenmovieticketnew/models/Movie.java
@@ -1,19 +1,20 @@
 package com.example.goldenmovieticketnew.models;
 
 import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.data.annotation.CreatedBy;
 import org.springframework.data.annotation.Id;
+import org.springframework.data.annotation.LastModifiedBy;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 import java.time.LocalDate;
 
 
 @Document(collection = "movies")
+@Data
+public class Movie   {
 
-public class Movie {
-    public String getId() {
-        return id;
-    }
     public Movie(){
 
     }
@@ -34,121 +35,6 @@ public class Movie {
         this.isShowing = isShowing;
     }
 
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getSmallImageURl() {
-        return smallImageURl;
-    }
-
-    public void setSmallImageURl(String smallImageURl) {
-        this.smallImageURl = smallImageURl;
-    }
-
-    public String getShortDescription() {
-        return shortDescription;
-    }
-
-    public void setShortDescription(String shortDescription) {
-        this.shortDescription = shortDescription;
-    }
-
-    public String getLongDescription() {
-        return longDescription;
-    }
-
-    public void setLongDescription(String longDescription) {
-        this.longDescription = longDescription;
-    }
-
-    public String getLargeImageURL() {
-        return largeImageURL;
-    }
-
-    public void setLargeImageURL(String largeImageURL) {
-        this.largeImageURL = largeImageURL;
-    }
-
-    public String getDirector() {
-        return director;
-    }
-
-    public void setDirector(String director) {
-        this.director = director;
-    }
-
-    public String getActors() {
-        return actors;
-    }
-
-    public void setActors(String actors) {
-        this.actors = actors;
-    }
-
-    public String getCategories() {
-        return categories;
-    }
-
-    public void setCategories(String categories) {
-        this.categories = categories;
-    }
-
-    public LocalDate getReleaseDate() {
-        return releaseDate;
-    }
-
-    public void setReleaseDate(LocalDate releaseDate) {
-        this.releaseDate = releaseDate;
-    }
-
-    public int getDuration() {
-        return duration;
-    }
-
-    public void setDuration(int duration) {
-        this.duration = duration;
-    }
-
-    public String getTrailerURL() {
-        return trailerURL;
-    }
-
-    public void setTrailerURL(String trailerURL) {
-        this.trailerURL = trailerURL;
-    }
-
-    public String getLanguage() {
-        return language;
-    }
-
-    public void setLanguage(String language) {
-        this.language = language;
-    }
-
-    public String getRated() {
-        return rated;
-    }
-
-    public void setRated(String rated) {
-        this.rated = rated;
-    }
-
-    public int getIsShowing() {
-        return isShowing;
-    }
-
-    public void setIsShowing(int isShowing) {
-        this.isShowing = isShowing;
-    }
 
     @Id
     private String id;
@@ -167,4 +53,11 @@ public class Movie {
     private String language;
     private String rated;
     private int isShowing;
+
+    private String createdDate;
+
+
+
+
+    private String lastModifiedDate;
 }
