+++ Golden Ticket New Back-End/goldenmovieticketnew/src/main/java/com/example/goldenmovieticketnew/models/Role.java
@@ -7,6 +7,7 @@ import org.springframework.data.mongodb.core.index.Indexed;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 @NoArgsConstructor
+@Data
 @Document(collection = "roles")
 public class Role {
 
@@ -15,11 +16,5 @@ public class Role {
     @Indexed(unique = true)
     private RoleName role;
 
-    public RoleName getRole() {
-        return role;
-    }
 
-    public void setRole(RoleName role) {
-        this.role = role;
-    }
 }
