+++ Golden Ticket New Back-End/goldenmovieticketnew/src/main/java/com/example/goldenmovieticketnew/models/User.java
@@ -12,7 +12,7 @@ import org.springframework.data.mongodb.core.mapping.Document;
 
 
 import java.util.Set;
-
+@Data
 @NoArgsConstructor
 @Document(collection = "users")
 public class User {
@@ -20,61 +20,6 @@ public class User {
     @Id
     private String id;
 
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getImage() {
-        return image;
-    }
-
-    public void setImage(String image) {
-        this.image = image;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getFullname() {
-        return fullname;
-    }
-
-    public void setFullname(String fullname) {
-        this.fullname = fullname;
-    }
-
-    public boolean isEnabled() {
-        return enabled;
-    }
-
-    public void setEnabled(boolean enabled) {
-        this.enabled = enabled;
-    }
-
-    public Set<Role> getRoles() {
-        return roles;
-    }
-
-    public void setRoles(Set<Role> roles) {
-        this.roles = roles;
-    }
 
     @Indexed(unique = true, direction = IndexDirection.DESCENDING)
     private String email;
