+++ Golden Ticket New Back-End/goldenmovieticketnew/src/main/java/com/example/goldenmovieticketnew/payload/LoginRequest.java
@@ -1,27 +1,17 @@
 package com.example.goldenmovieticketnew.payload;
 
-import javax.validation.constraints.NotBlank;
+import lombok.Data;
 
+import javax.validation.constraints.NotBlank;
+@Data
 public class LoginRequest {
 	@NotBlank
-	private String username;
+	private String usernameOrEmail;
 
 	@NotBlank
 	private String password;
 
-	public String getUsername() {
-		return username;
-	}
 
-	public void setUsername(String username) {
-		this.username = username;
-	}
 
-	public String getPassword() {
-		return password;
-	}
 
-	public void setPassword(String password) {
-		this.password = password;
-	}
 }
