+++ Golden Ticket New Back-End/goldenmovieticketnew/src/main/java/com/example/goldenmovieticketnew/payload/SignUpRequest.java
@@ -8,15 +8,15 @@ import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Size;
 import java.util.Set;
 
-
+@Data
 public class SignUpRequest {
 
     @NotBlank
-    @Size(min = 3, max = 20)
+
     private String username;
 
     @NotBlank
-    @Size(max = 50)
+
     @Email
     private String email;
 
@@ -28,55 +28,7 @@ public class SignUpRequest {
 
     private String image;
     private String fullname;
-    public String getUsername() {
-        return username;
-    }
 
-
-
-    public String getImage() {
-        return image;
-    }
-
-    public void setImage(String image) {
-        this.image = image;
-    }
-
-    public String getFullname() {
-        return fullname;
-    }
-
-    public void setFullname(String fullname) {
-        this.fullname = fullname;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public Set<String> getRoles() {
-        return this.roles;
-    }
-
-    public void setRole(Set<String> roles) {
-        this.roles = roles;
-    }
 
 
 }
