+++ Golden Ticket New Back-End/goldenmovieticketnew/src/main/java/com/example/goldenmovieticketnew/repositories/UserRepository.java
@@ -2,11 +2,14 @@ package com.example.goldenmovieticketnew.repositories;
 
 import com.example.goldenmovieticketnew.models.User;
 import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.data.mongodb.repository.Query;
 
 import java.util.Optional;
 
 public interface UserRepository extends MongoRepository<User,String> {
     Optional<User> findByEmail(String email);
+    @Query("{$or: [{'username': '?0'},{'email': '?0'}]}")
+    Optional<User> findByUsernameOrEmail(String username, String email);
     Optional<User> findByUsername(String username);
     Boolean existsByEmail(String email);
     Boolean existsByUsername(String username);
