+++ Golden Ticket New Back-End/goldenmovieticketnew/src/main/java/com/example/goldenmovieticketnew/secutiry/UserDetailsServiceImpl.java
@@ -20,9 +20,9 @@ public class UserDetailsServiceImpl implements UserDetailsService {
 
 	@Override
 	@Transactional
-	public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-		User user = userRepository.findByUsername(username)
-				.orElseThrow(() -> new UsernameNotFoundException("User Not Found with username: " + username));
+	public UserDetails loadUserByUsername(String usernameOrEmail) throws UsernameNotFoundException {
+		User user = userRepository.findByUsernameOrEmail(usernameOrEmail,usernameOrEmail)
+				.orElseThrow(() -> new UsernameNotFoundException("User Not Found with username or Email: " + usernameOrEmail));
 
 		return UserDetailsImpl.build(user);
 	}
