+++ Golden Ticket New Back-End/goldenmovieticketnew/src/main/resources/application.properties
@@ -11,3 +11,7 @@ spring.mvc.pathmatch.matching-strategy=ant_path_matcher
 # App Properties
 bezkoder.app.jwtSecret=9a02115a835ee03d5fb83cd8a468ea33e4090aaaec87f53c9fa54512bbef4db8dc656c82a315fa0c785c08b0
 bezkoder.app.jwtExpirationMs= 604800000
+
+## Jackson Properties
+spring.jackson.serialization.WRITE_DATES_AS_TIMESTAMPS= false
+spring.jackson.time-zone= UTC
